+++ analytical/templatetags/google_analytics.py
@@ -10,17 +10,13 @@ import re
 from django.conf import settings
 from django.template import Library, Node, TemplateSyntaxError
 
-from analytical.utils import is_internal_ip, disable_html, \
-        get_required_setting, get_domain, AnalyticalException
-
-
-def enumerate(sequence, start=0):
-    """Copy of the Python 2.6 `enumerate` builtin for compatibility."""
-    n = start
-    for elem in sequence:
-        yield n, elem
-        n += 1
-
+from analytical.utils import (
+    AnalyticalException,
+    disable_html,
+    get_domain,
+    get_required_setting,
+    is_internal_ip,
+)
 
 TRACK_SINGLE_DOMAIN = 1
 TRACK_MULTIPLE_SUBDOMAINS = 2
@@ -50,7 +46,7 @@ DOMAIN_CODE = "_gaq.push(['_setDomainName', '%s']);"
 NO_ALLOW_HASH_CODE = "_gaq.push(['_setAllowHash', false]);"
 ALLOW_LINKER_CODE = "_gaq.push(['_setAllowLinker', true]);"
 CUSTOM_VAR_CODE = "_gaq.push(['_setCustomVar', %(index)s, '%(name)s', " \
-        "'%(value)s', %(scope)s]);"
+                  "'%(value)s', %(scope)s]);"
 SITE_SPEED_CODE = "_gaq.push(['_trackPageLoadTime']);"
 ANONYMIZE_IP_CODE = "_gaq.push (['_gat._anonymizeIp']);"
 SAMPLE_RATE_CODE = "_gaq.push (['_gat._setSampleRate', '%s']);"
@@ -84,8 +80,8 @@ def google_analytics(parser, token):
 class GoogleAnalyticsNode(Node):
     def __init__(self):
         self.property_id = get_required_setting(
-                'GOOGLE_ANALYTICS_PROPERTY_ID', PROPERTY_ID_RE,
-                "must be a string looking like 'UA-XXXXXX-Y'")
+            'GOOGLE_ANALYTICS_PROPERTY_ID', PROPERTY_ID_RE,
+            "must be a string looking like 'UA-XXXXXX-Y'")
 
     def render(self, context):
         commands = self._get_domain_commands(context)
@@ -106,14 +102,15 @@ class GoogleAnalyticsNode(Node):
     def _get_domain_commands(self, context):
         commands = []
         tracking_type = getattr(settings, 'GOOGLE_ANALYTICS_TRACKING_STYLE',
-                TRACK_SINGLE_DOMAIN)
+                                TRACK_SINGLE_DOMAIN)
         if tracking_type == TRACK_SINGLE_DOMAIN:
             pass
         else:
             domain = get_domain(context, 'google_analytics')
             if domain is None:
-                raise AnalyticalException("tracking multiple domains with"
-                        " Google Analytics requires a domain name")
+                raise AnalyticalException(
+                    "tracking multiple domains with Google Analytics"
+                    " requires a domain name")
             commands.append(DOMAIN_CODE % domain)
             commands.append(NO_ALLOW_HASH_CODE)
             if tracking_type == TRACK_MULTIPLE_DOMAINS:
@@ -121,18 +118,24 @@ class GoogleAnalyticsNode(Node):
         return commands
 
     def _get_custom_var_commands(self, context):
-        values = (context.get('google_analytics_var%s' % i)
-                for i in range(1, 6))
-        vars = [(i, v) for i, v in enumerate(values, 1) if v is not None]
+        values = (
+            context.get('google_analytics_var%s' % i) for i in range(1, 6)
+        )
+        params = [(i, v) for i, v in enumerate(values, 1) if v is not None]
         commands = []
-        for index, var in vars:
+        for index, var in params:
             name = var[0]
             value = var[1]
             try:
                 scope = var[2]
             except IndexError:
                 scope = SCOPE_PAGE
-            commands.append(CUSTOM_VAR_CODE % locals())
+            commands.append(CUSTOM_VAR_CODE % {
+                'index': index,
+                'name': name,
+                'value': value,
+                'scope': scope,
+            })
         return commands
 
     def _get_other_commands(self, context):